refactor: splited build_app factory function from app instantiation
